pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "kintu"
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
]

[project.urls]
Homepage = "https://github.com/armandmcqueen/kintu"
Repository = "https://github.com/armandmcqueen/kintu.git"
Issues = "https://github.com/armandmcqueen/kintu/issues"

[tool.uv]
dev-dependencies = [
    "flit>=3.9.0",
    "invoke>=2.2.0",
    "twine>=6.1.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[tool.uv.sources]
kintu = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]  # E=errors, F=pyflakes, I=isort

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"
